moved BatchErrorResponse from the model package to an errorResponse package
