+++ src/main/java/com/revature/errorResponse/BatchErrorResponse.java
@@ -1,4 +1,4 @@
-package com.revature.models;
+package com.revature.errorResponse;
 
 public class BatchErrorResponse {
