+++ src/main/java/com/revature/repository/BatchRepository.java
@@ -11,7 +11,8 @@ import com.revature.models.Batch;
 
 /**
  * This is the repository that implements CRUD functionality directly to the Batch table in the database.  
- * The methods in this class are called from the BatchService class methods. You can create custom methods here, or use the built-in JPA methods available by extending the JPA repository such as findAll() and save().
+ * The methods in this class are called from the BatchService class methods. You can create custom methods here, 
+ * or use the built-in JPA methods available by extending the JPA repository such as findAll() and save().
  * @author John Beineke, Batch: 1806-Jun18-Java-USF, Trainer: Wezley Singleton
  */
